+++ SDLGame-TFG/GraphicEngine/matrixscript.cpp
@@ -54,7 +54,7 @@ Matrix4f & MatrixScript::updateState(float time){
     if(currentTime>script[currentMove].first){
         currentTime=0;
         currentMove++;
-        if(currentMove==script.size()){ //initial Case
+        if((unsigned)currentMove==script.size()){ //initial Case
             currentMove=0;
         }
         (script[currentMove].second)->resetState(); //Reset the state
@@ -67,7 +67,7 @@ Matrix4f & MatrixScript::updateState(float time){
 
 void MatrixScript::resetState(){
     currentMove=0;
-    for(int i=0;i<script.size();i++)
+    for(unsigned i=0;i<script.size();i++)
         (script[i].second)->resetState();
 }
